pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.13.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.2.0",
    "langchain-text-splitters>=1.1.2",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
]
